source/web-django/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pillow==10.2.0
gunicorn==23.0.0
python-decouple==3.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
uuid >= 1.30
requests

